+++ code/src/cljs/sixsq/nuvla/ui/profile/views.cljs
@@ -1124,7 +1124,8 @@
         customer-info (subscribe [::subs/customer-info])]
     (fn []
       (let [{:keys [street-address city country postal-code]} (:address @customer-info)
-            {:keys [fullname balance currency] :or {balance 0}} @customer-info]
+            {:keys [fullname balance currency] :or {balance 0}} @customer-info
+            balance (* balance -1)]
         [ui/Segment {:padded  true
                      :color   "grey"
                      :loading @loading?
@@ -1138,9 +1139,10 @@
            [ui/TableRow
             [ui/TableCell [:b (@tr [:street-address])]]
             [ui/TableCell street-address [:br] postal-code " - " city [:br] country]]
-          [ui/TableRow
-           [ui/TableCell {:width 5} [:b (str/capitalize (@tr [:balance]))]]
-           [ui/TableCell {:width 11} (format-currency currency balance)]]]]]))))
+           [ui/TableRow {:negative (neg? balance)
+                         :positive (pos? balance)}
+            [ui/TableCell {:width 5} [:b (str/capitalize (@tr [:balance]))]]
+            [ui/TableCell {:width 11} (format-currency currency balance)]]]]]))))
 
 
 (defn DropdownPrincipals